# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AssetRecord": {
        "properties": {
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bytes"
          },
          "collection_ids": {
            "items": {
              "type": "string"
            },
            "title": "Collection Ids",
            "type": "array"
          },
          "compiled_graph_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compiled Graph Hash"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "duplicate_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Of"
          },
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "fps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fps"
          },
          "frames": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frames"
          },
          "h": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "H"
          },
          "has_document": {
            "default": false,
            "title": "Has Document",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "kind": {
            "default": "image",
            "enum": [
              "image",
              "video",
              "mask",
              "document-render"
            ],
            "title": "Kind",
            "type": "string"
          },
          "lineage_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lineage Kind"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "parents": {
            "items": {
              "type": "string"
            },
            "title": "Parents",
            "type": "array"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "prompt_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Json"
          },
          "prompt_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Text"
          },
          "rating": {
            "default": 0,
            "title": "Rating",
            "type": "integer"
          },
          "root_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Id"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "state": {
            "default": "none",
            "enum": [
              "none",
              "keep",
              "reject"
            ],
            "title": "State",
            "type": "string"
          },
          "suite": {
            "default": "generate",
            "title": "Suite",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "thumb_status": {
            "default": "pending",
            "enum": [
              "pending",
              "done",
              "failed"
            ],
            "title": "Thumb Status",
            "type": "string"
          },
          "timings": {
            "additionalProperties": true,
            "title": "Timings",
            "type": "object"
          },
          "trashed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trashed At"
          },
          "variant": {
            "default": "full",
            "title": "Variant",
            "type": "string"
          },
          "w": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "W"
          }
        },
        "required": [
          "schema_version",
          "id",
          "kind",
          "path",
          "w",
          "h",
          "frames",
          "fps",
          "created_at",
          "job_id",
          "batch_id",
          "session_id",
          "root_id",
          "parents",
          "lineage_kind",
          "suite",
          "model_id",
          "format",
          "seed",
          "prompt_text",
          "prompt_json",
          "params",
          "timings",
          "compiled_graph_hash",
          "variant",
          "state",
          "rating",
          "tags",
          "collection_ids",
          "duplicate_of",
          "has_document",
          "thumb_status",
          "trashed_at",
          "bytes",
          "sha256"
        ],
        "title": "AssetRecord",
        "type": "object"
      },
...
      "BlobPut": {
        "properties": {
          "bytes": {
            "title": "Bytes",
            "type": "integer"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "sha256",
          "bytes"
        ],
        "title": "BlobPut",
        "type": "object"
      },
...
      "ClipExtract": {
        "properties": {
          "frames": {
            "items": {
              "type": "integer"
            },
            "title": "Frames",
            "type": "array"
          }
        },
        "required": [
          "frames"
        ],
        "title": "ClipExtract",
        "type": "object"
      },
      "ClipExtracted": {
        "properties": {
          "clip": {
            "$ref": "#/components/schemas/ClipRecord"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AssetRecord"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "clip"
        ],
        "title": "ClipExtracted",
        "type": "object"
      },
...
      "ClipRecord": {
        "properties": {
          "asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "beats": {
            "items": {
              "properties": {
                "asset_id": {
                  "type": "string"
                },
                "frame": {
                  "type": "integer"
                },
                "strength": {
                  "type": "number"
                }
              },
              "required": [
                "frame",
                "asset_id",
                "strength"
              ],
              "title": "ClipBeat",
              "type": "object"
            },
            "title": "Beats",
            "type": "array"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "end_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Asset Id"
          },
          "extracted_asset_ids": {
            "items": {
              "type": "string"
            },
            "title": "Extracted Asset Ids",
            "type": "array"
          },
          "fps": {
            "title": "Fps",
            "type": "integer"
          },
          "frames": {
            "title": "Frames",
            "type": "integer"
          },
          "h": {
            "title": "H",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "identity": {
            "anyOf": [
              {
                "additionalProperties": true,
                "properties": {
                  "mean": {
                    "type": "number"
                  },
                  "min": {
                    "type": "number"
                  },
                  "min_frame": {
                    "type": "integer"
                  },
                  "model": {
                    "type": "string"
                  },
                  "per_frame": {
                    "items": {
                      "properties": {
                        "faces": {
                          "type": "integer"
                        },
                        "frame": {
                          "type": "integer"
                        },
                        "sim": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      },
                      "required": [
                        "frame",
                        "sim",
                        "faces"
                      ],
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "reference_faces": {
                    "type": "integer"
                  },
                  "sampled": {
                    "type": "integer"
                  },
                  "scale": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "with_face": {
                    "type": "integer"
                  }
                },
                "required": [
                  "status"
                ],
                "title": "ClipIdentity",
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identity"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "master_dir": {
            "title": "Master Dir",
            "type": "string"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "preset": {
            "default": "draft",
            "title": "Preset",
            "type": "string"
          },
          "prompt": {
            "default": "",
            "title": "Prompt",
            "type": "string"
          },
          "proxy_bytes": {
            "default": 0,
            "title": "Proxy Bytes",
            "type": "integer"
          },
          "proxy_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Path"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "seed": {
            "default": 0,
            "title": "Seed",
            "type": "integer"
          },
          "start_asset_id": {
            "title": "Start Asset Id",
            "type": "string"
          },
          "timings": {
            "additionalProperties": true,
            "title": "Timings",
            "type": "object"
          },
          "w": {
            "title": "W",
            "type": "integer"
          }
        },
        "required": [
          "schema_version",
          "id",
          "created_at",
          "job_id",
          "batch_id",
          "asset_id",
          "model_id",
          "preset",
          "prompt",
          "seed",
          "frames",
          "fps",
          "w",
          "h",
          "start_asset_id",
          "end_asset_id",
          "beats",
          "master_dir",
          "proxy_path",
          "proxy_bytes",
          "extracted_asset_ids",
          "params",
          "timings",
          "identity"
        ],
        "title": "ClipRecord",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "loom2 orchestrator",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/blobs/{sha}": {
      "get": {
        "operationId": "blob_get_blobs__sha__get",
        "parameters": [
          {
            "in": "path",
            "name": "sha",
            "required": true,
            "schema": {
              "title": "Sha",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Blob Get",
        "tags": [
          "blobs"
        ]
      },
      "put": {
        "operationId": "blob_put_blobs__sha__put",
        "parameters": [
          {
            "in": "path",
            "name": "sha",
            "required": true,
            "schema": {
              "title": "Sha",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlobPut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Blob Put",
        "tags": [
          "blobs"
        ]
      }
    },
...
    "/clips/{clip_id}": {
      "get": {
        "operationId": "clip_get_clips__clip_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "clip_id",
            "required": true,
            "schema": {
              "title": "Clip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clip Get",
        "tags": [
          "clips"
        ]
      }
    },
    "/clips/{clip_id}/extract": {
      "post": {
        "description": "Master frames → Catalogue images with `frame-extract` lineage from the clip's asset (11 §6).",
        "operationId": "clip_extract_clips__clip_id__extract_post",
        "parameters": [
          {
            "in": "path",
            "name": "clip_id",
            "required": true,
            "schema": {
              "title": "Clip Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClipExtract"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipExtracted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clip Extract",
        "tags": [
          "clips"
        ]
      }
    },
    "/clips/{clip_id}/frames/{name}": {
      "get": {
        "description": "A master frame as PNG; with `size` a cached WebP thumbnail (longer side ≤ size) for the filmstrip (C24).",
        "operationId": "clip_frame_clips__clip_id__frames__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "clip_id",
            "required": true,
            "schema": {
              "title": "Clip Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clip Frame",
        "tags": [
          "clips"
        ]
      }
    },
    "/clips/{clip_id}/identity": {
      "post": {
        "description": "FaceSim advisory on demand (11 §6): cosine similarity of the start frame's face across the clip.",
        "operationId": "clip_identity_clips__clip_id__identity_post",
        "parameters": [
          {
            "in": "path",
            "name": "clip_id",
            "required": true,
            "schema": {
              "title": "Clip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clip Identity",
        "tags": [
          "clips"
        ]
      }
    },
    "/clips/{clip_id}/proxy.mp4": {
      "get": {
        "operationId": "clip_proxy_clips__clip_id__proxy_mp4_get",
        "parameters": [
          {
            "in": "path",
            "name": "clip_id",
            "required": true,
            "schema": {
              "title": "Clip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clip Proxy",
        "tags": [
          "clips"
        ]
      }
    },
...
  }
}